chore(message-list): split message list screen composable previews in smaller previews
